Change percentage to float instead of int

+++ server/models/measurement.js
@@ -17,7 +17,7 @@ module.exports = (sequelize, DataTypes) => {
     sensorID: DataTypes.INTEGER,
     plantID: DataTypes.INTEGER,
     moistureValue: DataTypes.INTEGER,
-    moisturePercentage: DataTypes.INTEGER,
+    moisturePercentage: DataTypes.FLOAT,
   }, {
     sequelize,
     modelName: 'Measurement',
